refactor: sqliteProxy
